Conflicts:
	TFC API/TFC/API/Constant/TFCBlockID.java
	TFC_Shared/src/TFC/Items/ItemMetalSheet.java
	TFC_Shared/src/TFC/Items/ItemMetalSheet2x.java
	TFC_Shared/src/TFC/Items/Tools/ItemCustomBow.java
	TFC_Shared/src/TFC/Items/Tools/ItemJavelin.java
	TFC_Shared/src/TFC/Items/Tools/ItemMiscToolHead.java
	TFC_Shared/src/TFC/TerraFirmaCraft.java
	TFC_Shared/src/TFC/TileEntities/TileEntityFirepit.java

+++ TFC_Shared/src/TFC/Core/Metal/MetalPair.java
@@ -5,8 +5,8 @@ import TFC.API.Metal;
 public class MetalPair 
 {
 	public Metal type;
-	public short amount;
-	public MetalPair(Metal t, short amnt)
+	public float amount;
+	public MetalPair(Metal t, float amnt)
 	{
 		type = t;
 		amount = amnt;
